done crud

+++ index.php
@@ -6,7 +6,7 @@
 <body>
 
 	<h2>Data Mahasiswa</h2>
-	<p><a href="">Tambah Data</a></p>
+	<p><a href="tambah.php">Tambah Data</a></p>
 	<table cellpadding="5" cellspacing="0" border="1">
 		<tr bgcolor="#ddd">
 			<th>No.</th>
@@ -19,7 +19,7 @@
 		</tr>
 		<?php
 			include 'koneksi.php';
-			$query = mysqli_query($koneksi,"SELECT * FROM bio ORDER BY id_mhs DESC") or die(mysqli_error($koneksi));
+			$query = mysqli_query($koneksi,"SELECT * FROM bio ORDER BY id_mhs") or die(mysqli_error($koneksi));
 			if(mysqli_num_rows($query)==0){ ?>
 				<tr>
 					<td colspan="8">Tidak ada data</td>
@@ -28,7 +28,7 @@
 			}
 			else{
 				$no = 1;
-				while($data = mysql_fetch_assoc($query)){
+				while($data = mysqli_fetch_assoc($query)){
 				?>
 				<tr>
 					<td><?php echo $no?></td>
@@ -37,7 +37,7 @@
 					<td><?php echo $data['jk']?></td>
 					<td><?php echo $data['alamat']?></td>
 					<td><?php echo $data['no_hp']?></td>
-					<td><a href="edit.php?id=<?php echo $data['id_mhs'] ?>">Edit</a> <a href="hapus.php?id=<?php echo $data['id_mhs'] ?>">Hapus</a></td>
+					<td><a href="edit.php?id=<?php echo $data['id_mhs']; ?>">Edit</a> <a href="hapus.php?id=<?php echo $data['id_mhs']; ?>" onclick="return confirm('Yakin ingin menghapus?')">Hapus</a></td>
 				</tr>
 				<?php
 				$no++;
